+++ ProkenB/Assets/Scenes/3c1u/_GameManager.unity
@@ -297,7 +297,7 @@ RectTransform:
   m_Children:
   - {fileID: 1800382122}
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -381,7 +381,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1800382121
 GameObject:
@@ -567,85 +567,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 50, y: -30, z: 0}
---- !u!1001 &6233999273375945452
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 3
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283404, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283440, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: m_Name
-      value: Game
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283443, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: viewIdField
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 6233999273046283443, guid: 8e8dcf1a6097d4fb38aca2bf44003440,
-        type: 3}
-      propertyPath: OwnershipTransfer
-      value: 1
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 8e8dcf1a6097d4fb38aca2bf44003440, type: 3}
 --- !u!1001 &7795579380237011200
 PrefabInstance:
   m_ObjectHideFlags: 0
