--- ProkenB/Assets/Scripts/ProkenB/Networking.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5fd86f1440bd24f2481109f4de466b81
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
