--- ProkenB/Assets/Scripts/ProkenB/Networking/NetworkManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 328b9fd5c508d4238abe0f7510814e53
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
