--- ProkenB/Assets/Scripts/ProkenB/View/GameView.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 5df9c0c19df249cd9bf29891bbaf868d
-timeCreated: 1603686505
